requirements.txt
# Core Framework
langgraph>=0.2.0
langchain>=0.3.0

# Google Cloud / Vertex AI
langchain-google-genai>=2.0.0
langchain-google-vertexai>=2.0.0
google-cloud-aiplatform>=1.60.0
google-cloud-firestore>=2.16.0
google-auth>=2.0.0

# Data Validation & Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
pandas>=2.0.0

# API Server
fastapi>=0.115.0
uvicorn[standard]>=0.27.0

# Database Connectors
pymssql>=2.2.0
